Fix ec_goals test on Chrome
Was failing because Chrome includes the accessible title text in these
droplet blocks when you ask for the text-content of the top-level SVG
group, where Phantom was only including the visible text.

Fixed by making the selector more specific, only grabbing the subset of
visible text that changes from block to block, and updating the
expectation accordingly.